fix(auth): rename oauth files and README wording
